refactor: organize keymaps into dedicated modules
