+++ lua/core/keymaps/init.lua
@@ -1,0 +1,4 @@
+-- load all keymap modules
+require("core.keymaps.general").setup()
+
+return {}
