+++ lua/plugins/navigation.lua
@@ -7,17 +7,8 @@ return {
       require("smart-splits").setup()
       local ss = require("smart-splits")
 
-      -- 使用 Ctrl+方向键在窗口之间切换
-      vim.keymap.set("n", "<C-h>", ss.move_cursor_left)
-      vim.keymap.set("n", "<C-j>", ss.move_cursor_down)
-      vim.keymap.set("n", "<C-k>", ss.move_cursor_up)
-      vim.keymap.set("n", "<C-l>", ss.move_cursor_right)
-
-      -- 使用 Alt+方向键调整当前窗口大小
-      vim.keymap.set("n", "<A-h>", ss.resize_left)
-      vim.keymap.set("n", "<A-j>", ss.resize_down)
-      vim.keymap.set("n", "<A-k>", ss.resize_up)
-      vim.keymap.set("n", "<A-l>", ss.resize_right)
+      -- 键位绑定统一管理
+      require("core.keymaps.smart_splits").setup(ss)
     end,
   },
 
@@ -27,10 +18,7 @@ return {
   "nvim-telescope/telescope.nvim",
   dependencies = { "nvim-lua/plenary.nvim" },
 
-  keys = {
-    { "<leader>ff", "<cmd>Telescope find_files<CR>", desc = "查找文件" },
-    { "<leader>fg", "<cmd>Telescope live_grep<CR>",  desc = "全局搜索" },
-  },
+  keys = require("core.keymaps.telescope").keys,
 
   cmd  = "Telescope",       -- 可留可删
   opts = true,              -- 若只是默认配置，留 true 即可；若自定义请写表
